:wrench: switch contact components to ts
