+++ Components/GeneralComponents/ColoredText.tsx
@@ -2,7 +2,7 @@ import React,{FC, ReactNode} from 'react'
 interface Props{
   children:ReactNode,
   color:string,
-  className:string
+  className?:string
 }
 const ColoredText:FC<Props> = ({children, color, className, ...rest}) => {
   return (
