+++ pages/about-me/hobbies.jsx
@@ -39,7 +39,6 @@ const Hobbies = ({files}) => {
         <article className="w-full  border-t-lines lg:w-[41.4%] h-fit lg:h-full text-secondary text-xs ">
             <LeftContent  openEditor={openEditor} handleChangeEditor={handleChangeEditor} closeEditor={closeEditor}/>
         </article>
-   
         <div className="hidden lg:block w-5 h-full mt-8 px-1 border-t-2 border-t-lines">
               <div className="w-3 h-full mx-auto mt-1 border-t-[6px] border-t-secondary">
               </div>
